RAG/GenAI-RAG-Context-Aware-Customer-Assistant-for-E-Commerce-Organic-Farm-to-Table-Store-main/requirements.txt
flask==2.0.1
requests>=2.28.0
chromadb==0.3.21
sentence-transformers==2.2.2
langchain==0.0.184
openai==0.27.0
python-dotenv==0.19.2
werkzeug==2.0.3
